requirements.txt
streamlit
pdfplumber
PyMuPDF
python-dotenv
google-genai
requests
